Make Data Access Hatch an universal hatch
